backend/requirements/base.txt
# Core Django Framework
Django>=5.0,<5.2
djangorestframework>=3.15,<4.0

# Multi-tenancy Support  
django-tenants>=3.6,<4.0

# Database (PostgreSQL)
psycopg2-binary>=2.9,<3.0

# Authentication & Security
PyJWT>=2.8,<3.0
django-cors-headers>=4.3,<5.0

# Configuration Management
python-dotenv>=1.0,<2.0

# File Upload Support
Pillow>=10.3,<11.0

# API Documentation
drf-spectacular>=0.27,<0.28